ArchLinux/gui-installer/requirements.txt
tkinter
pillow
ttkthemes
ttkbootstrap>=1.0.0